requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
fiona
shapely
rtree
